[149] Déplacer les actions sur les groupes dans le GroupController
