+++ app/controllers/UserController.scala
@@ -3,16 +3,14 @@ package controllers
 import java.util.{Locale, UUID}
 
 import javax.inject.{Inject, Singleton}
-import actions.{LoginAction, RequestWithUserData}
+import actions.LoginAction
 import extentions.Operators.{GroupOperators, UserOperators, not}
-import extentions.{Time, UUIDHelper}
+import extentions.UUIDHelper
+import forms.Models
 import models.{Area, User, UserGroup}
 import org.joda.time.{DateTime, DateTimeZone}
 import org.postgresql.util.PSQLException
 import org.webjars.play.WebJarsUtil
-import play.api.data.Form
-import play.api.data.Forms._
-import play.api.data.validation.Constraints._
 import play.api.mvc.{Action, AnyContent, Call, InjectedController}
 import play.filters.csrf.CSRF
 import play.filters.csrf.CSRF.Token
@@ -25,6 +23,8 @@ case class UserController @Inject()(loginAction: LoginAction,
                                     applicationService: ApplicationService,
                                     notificationsService: NotificationService,
                                     eventService: EventService)(implicit val webJarsUtil: WebJarsUtil) extends InjectedController with play.api.i18n.I18nSupport with UserOperators with GroupOperators {
+
+  private implicit val timeZone :DateTimeZone= DateTimeZone.forID("Europe/Paris")
 
   def all(areaId: UUID) = loginAction { implicit request =>
     if(request.currentUser.canSeeUsersInArea(areaId) == false) {
@@ -56,7 +56,6 @@ case class UserController @Inject()(loginAction: LoginAction,
       Ok(views.html.allUsers(request.currentUser)(groups, users, applications, selectedArea))
     }
   }
-
 
   def allCSV(areaId: java.util.UUID) = loginAction { implicit request =>
     if(request.currentUser.admin == false || request.currentUser.canSeeUsersInArea(areaId) == false) {
@@ -100,67 +99,6 @@ case class UserController @Inject()(loginAction: LoginAction,
     }
   }
 
-  private val timeZone = DateTimeZone.forID("Europe/Paris")
-
-  def userMapping = mapping(
-    "id" -> optional(uuid).transform[UUID]({
-        case None => UUID.randomUUID()
-        case Some(id) => id
-      }, { Some(_) }),
-    "key" -> ignored("key"),
-    "name" -> nonEmptyText.verifying(maxLength(100)),
-    "qualite" -> nonEmptyText.verifying(maxLength(100)),
-    "email" -> email.verifying(maxLength(200), nonEmpty),
-    "helper" -> boolean,
-    "instructor" -> boolean,
-    "admin" -> boolean,
-    "areas" -> list(uuid).verifying("Vous devez sélectionner au moins un territoire", _.nonEmpty),
-    "creationDate" -> ignored(DateTime.now(timeZone)),
-    "hasAcceptedCharte" -> boolean,
-    "communeCode" -> default(nonEmptyText.verifying(maxLength(5)), "0"),
-    "adminGroup" -> boolean,
-    "disabled" -> boolean,
-    "expert" -> ignored(false),
-    "groupIds" -> default(list(uuid), List()),
-    "delegations" -> seq(tuple(
-      "name" -> nonEmptyText,
-      "email" -> email
-    )).transform[Map[String,String]]({ _.toMap }, { _.toSeq }),
-    "cguAcceptationDate" -> optional(ignored(Time.now())),
-    "newsletterAcceptationDate" -> optional(ignored(Time.now()))
-  )(User.apply)(User.unapply)
-
-  val userForm = Form(userMapping)
-
-  def usersForm(implicit area: Area) = Form(
-    single(
-      "users" -> list(mapping(
-        "id" -> optional(uuid).transform[UUID]({
-          case None => UUID.randomUUID()
-          case Some(id) => id
-        }, { Some(_) }),
-        "key" -> ignored("key"),
-        "name" -> nonEmptyText.verifying(maxLength(100)),
-        "qualite" -> nonEmptyText.verifying(maxLength(100)),
-        "email" -> email.verifying(maxLength(200), nonEmpty),
-        "helper" -> boolean,
-        "instructor" -> boolean,
-        "admin" -> ignored(false),
-        "areas" -> ignored(List(area.id)),
-        "creationDate" -> ignored(DateTime.now(timeZone)),
-        "hasAcceptedCharte" -> ignored(false),
-        "communeCode" -> default(nonEmptyText.verifying(maxLength(5)), "0"),
-        "adminGroup" -> ignored(false),
-        "disabled" -> ignored(false),
-        "expert" -> ignored(false),
-        "groupIds" -> default(list(uuid), List()),
-        "delegations" -> ignored(Map[String,String]()),
-        "cguAcceptationDate" -> optional(ignored(Time.now())),
-        "newsletterAcceptationDate" -> optional(ignored(Time.now()))
-      )(User.apply)(User.unapply))
-    )
-  )
-
   def editUser(userId: UUID): Action[AnyContent] = loginAction { implicit request =>
     asAdmin { () =>
       "VIEW_USER_UNAUTHORIZED" -> s"Accès non autorisé pour voir $userId"
@@ -170,7 +108,7 @@ case class UserController @Inject()(loginAction: LoginAction,
           eventService.error("USER_NOT_FOUND", s"L'utilisateur $userId n'existe pas")
           NotFound("Nous n'avons pas trouvé cet utilisateur")
         case Some(user) if user.canBeEditedBy(request.currentUser) =>
-          val form = userForm.fill(user)
+          val form = Models.userForm.fill(user)
           val groups = groupService.allGroups
           val unused = not(isAccountUsed(user))
           val Token(tokenName, tokenValue) = CSRF.getToken.get
@@ -208,7 +146,7 @@ case class UserController @Inject()(loginAction: LoginAction,
     asAdmin { () =>
       "POST_EDIT_USER_UNAUTHORIZED" -> s"Accès non autorisé à modifier $userId"
     } { () =>
-      userForm.bindFromRequest.fold(
+      Models.userForm.bindFromRequest.fold(
         formWithErrors => {
           val groups = groupService.allGroups
           eventService.error("ADD_USER_ERROR", s"Essai de modification de l'tilisateur $userId avec des erreurs de validation")
@@ -222,7 +160,7 @@ case class UserController @Inject()(loginAction: LoginAction,
               eventService.info("EDIT_USER_DONE", s"Utilisateur $userId modifié", user = Some(updatedUser))
               Redirect(routes.UserController.all(Area.allArea.id)).flashing("success" -> "Utilisateur modifié")
             } else {
-              val form = userForm.fill(updatedUser).withGlobalError("Impossible de mettre à jour l'utilisateur $userId (Erreur interne)")
+              val form = Models.userForm.fill(updatedUser).withGlobalError("Impossible de mettre à jour l'utilisateur $userId (Erreur interne)")
               val groups = groupService.allGroups
               eventService.error("EDIT_USER_ERROR", s"Impossible de modifier l'utilisateur dans la BDD", user = Some(updatedUser))
               InternalServerError(views.html.editUser(request.currentUser, request.currentArea)(form, userId, groups))
@@ -230,20 +168,6 @@ case class UserController @Inject()(loginAction: LoginAction,
           }
         }
       )
-    }
-  }
-
-  def add(groupId: UUID): Action[AnyContent] = loginAction { implicit request =>
-    withGroup(groupId) { group: UserGroup =>
-      if (!group.canHaveUsersAddedBy(request.currentUser)) {
-        eventService.warn("SHOW_ADD_USER_UNAUTHORIZED", s"Accès non autorisé à l'admin des utilisateurs du groupe $groupId")
-        Unauthorized("Vous n'avez pas le droit de faire ça")
-      } else {
-        implicit val area = Area.fromId(group.area).get
-        val rows = request.getQueryString("rows").map(_.toInt).getOrElse(1)
-        eventService.info("EDIT_USER_SHOWED", s"Visualise la vue d'ajouts des utilisateurs")
-        Ok(views.html.editUsers(request.currentUser, request.currentArea)(usersForm, rows, routes.UserController.addPost(groupId)))
-      }
     }
   }
 
@@ -255,7 +179,7 @@ case class UserController @Inject()(loginAction: LoginAction,
       } else {
         val group = groupService.groupById(groupId).get
         implicit val area = Area.fromId(group.area).get
-        usersForm.bindFromRequest.fold(
+        Models.usersForm.bindFromRequest.fold(
           formWithErrors => {
             eventService.error("ADD_USER_ERROR", s"Essai d'ajout d'utilisateurs avec des erreurs de validation")
             BadRequest(views.html.editUsers(request.currentUser, request.currentArea)(formWithErrors, 0, routes.UserController.addPost(groupId)))
@@ -264,9 +188,9 @@ case class UserController @Inject()(loginAction: LoginAction,
             try {
               if (userService.add(users.map(_.copy(groupIds = List(groupId))))) {
                 eventService.info("ADD_USER_DONE", s"Utilisateurs ajouté")
-                Redirect(routes.UserController.editGroup(groupId)).flashing("success" -> "Utilisateurs ajouté")
+                Redirect(routes.GroupController.editGroup(groupId)).flashing("success" -> "Utilisateurs ajouté")
               } else {
-                val form = usersForm.fill(users).withGlobalError("Impossible d'ajouté les utilisateurs (Erreur interne 1)")
+                val form = Models.usersForm.fill(users).withGlobalError("Impossible d'ajouté les utilisateurs (Erreur interne 1)")
                 eventService.error("ADD_USER_ERROR", s"Impossible d'ajouter des utilisateurs dans la BDD 1")
                 InternalServerError(views.html.editUsers(request.currentUser, request.currentArea)(form, users.length, routes.UserController.addPost(groupId)))
               }
@@ -277,7 +201,7 @@ case class UserController @Inject()(loginAction: LoginAction,
                   case Some(email) => s"Un utilisateur avec l'adresse $email existe déjà."
                   case _ => "Erreur d'insertion dans la base de donnée : contacter l'administrateur."
                 }
-                val form = usersForm.fill(users).withGlobalError(errorMessage)
+                val form = Models.usersForm.fill(users).withGlobalError(errorMessage)
                 eventService.error("ADD_USER_ERROR", s"Impossible d'ajouter des utilisateurs dans la BDD : ${ex.getServerErrorMessage}")
                 BadRequest(views.html.editUsers(request.currentUser, request.currentArea)(form, users.length, routes.UserController.addPost(groupId)))
             }
@@ -292,16 +216,8 @@ case class UserController @Inject()(loginAction: LoginAction,
     Ok(views.html.showCGU(request.currentUser, request.currentArea))
   }
 
-  private val validateCGUForm = Form(
-    tuple(
-      "redirect" -> optional(text),
-      "newsletter" -> boolean,
-      "validate" -> boolean
-    )
-  )
-
   def validateCGU(): Action[AnyContent] = loginAction { implicit request =>
-    validateCGUForm.bindFromRequest.fold(
+    Models.validateCGUForm.bindFromRequest.fold(
       formWithErrors => {
         eventService.error("CGU_VALIDATION_ERROR", s"Erreur de formulaire dans la validation des CGU")
         BadRequest(s"Formulaire invalide, prévenez l'administrateur du service. ${formWithErrors.errors.mkString(", ")}")
@@ -321,75 +237,19 @@ case class UserController @Inject()(loginAction: LoginAction,
         }
       }
     )
-
   }
 
-  def editGroup(id: UUID): Action[AnyContent] = loginAction { implicit request =>
-    withGroup(id) { group: UserGroup =>
+  def add(groupId: UUID): Action[AnyContent] = loginAction { implicit request =>
+    withGroup(groupId) { group: UserGroup =>
       if (!group.canHaveUsersAddedBy(request.currentUser)) {
-        eventService.warn("EDIT_GROUPE_UNAUTHORIZED", s"Accès non autorisé à l'edition de ce groupe")
-        Unauthorized("Vous ne pouvez pas éditer ce groupe : êtes-vous dans la bonne zone ?")
+        eventService.warn("SHOW_ADD_USER_UNAUTHORIZED", s"Accès non autorisé à l'admin des utilisateurs du groupe $groupId")
+        Unauthorized("Vous n'avez pas le droit de faire ça")
       } else {
-        val groupUsers = userService.byGroupIds(List(id))
-        eventService.info("EDIT_GROUP_SHOWED", s"Visualise la vue de modification du groupe")
-        val isEmpty = groupService.isGroupEmpty(group.id)
-        Ok(views.html.editGroup(request.currentUser, request.currentArea)(group, groupUsers, isEmpty))
+        implicit val area = Area.fromId(group.area).get
+        val rows = request.getQueryString("rows").map(_.toInt).getOrElse(1)
+        eventService.info("EDIT_USER_SHOWED", s"Visualise la vue d'ajouts des utilisateurs")
+        Ok(views.html.editUsers(request.currentUser, request.currentArea)(Models.usersForm, rows, routes.UserController.addPost(groupId)))
       }
-    }
-  }
-
-  def editGroupPost(id: UUID): Action[AnyContent] = loginAction { implicit request =>
-    asAdmin { () =>
-      "EDIT_GROUPE_UNAUTHORIZED" -> s"Accès non autorisé à l'edition de ce groupe"
-    } { () =>
-      withGroup(id) { _: UserGroup =>
-        addGroupForm.bindFromRequest.fold(_ => {
-          eventService.error("EDIT_USER_GROUP_ERROR", s"Essai d'edition d'un groupe avec des erreurs de validation")
-          BadRequest("Impossible de modifier le groupe (erreur de formulaire)")
-        }, group => {
-          if (groupService.edit(group.copy(id = id))) {
-            eventService.info("EDIT_USER_GROUP_DONE", s"Groupe édité")
-            Redirect(routes.UserController.editGroup(id)).flashing("success" -> "Groupe modifié")
-          } else {
-            eventService.error("EDIT_USER_GROUP_ERROR", s"Impossible de modifier le groupe dans la BDD")
-            Redirect(routes.UserController.editGroup(id)).flashing("success" -> "Impossible de modifier le groupe")
-          }
-        }
-        )
-      }
-    }
-  }
-
-  def addGroupForm[A](implicit request: RequestWithUserData[A]) = Form(
-    mapping(
-      "id" -> ignored(UUID.randomUUID()),
-      "name" -> text(maxLength = 50),
-      "description" -> optional(text),
-      "insee-code" -> text,
-      "creationDate" -> ignored(DateTime.now(timeZone)),
-      "create-by-user-id" -> ignored(request.currentUser.id),
-      "area" -> uuid.verifying("Vous devez sélectionner un territoire sur lequel vous êtes admin", area => request.currentUser.areas.contains(area)),
-      "organisation" -> optional(text),
-      "email" -> optional(email)
-    )(UserGroup.apply)(UserGroup.unapply)
-  )
-
-  def addGroup(): Action[AnyContent] = loginAction { implicit request =>
-    asAdmin { () =>
-      "ADD_GROUP_UNAUTHORIZED" -> s"Accès non autorisé pour ajouter un groupe"
-    } { () =>
-      addGroupForm.bindFromRequest.fold(_ => {
-        eventService.error("ADD_USER_GROUP_ERROR", s"Essai d'ajout d'un groupe avec des erreurs de validation")
-        BadRequest("Impossible d'ajouter le groupe") //BadRequest(views.html.editUsers(request.currentUser, request.currentArea)(formWithErrors, 0, routes.UserController.addPost()))
-      }, group => {
-        if (groupService.add(group)) {
-          eventService.info("ADD_USER_GROUP_DONE", s"Groupe ajouté")
-          Redirect(routes.UserController.editGroup(group.id)).flashing("success" -> "Groupe ajouté")
-        } else {
-          eventService.error("ADD_USER_GROUP_ERROR", s"Impossible d'ajouter le groupe dans la BDD")
-          Redirect(routes.UserController.all(Area.allArea.id)).flashing("success" -> "Impossible d'ajouter le groupe")
-        }
-      })
     }
   }
