+++ conf/routes
@@ -40,9 +40,9 @@ GET     /user/delete/unused/:userId  controllers.UserController.deleteUnusedUser
 
 
 # Groups
-POST    /groups                         controllers.UserController.addGroup
-GET     /groups/:groupId                controllers.UserController.editGroup(groupId: java.util.UUID)
-POST    /groups/:groupId                controllers.UserController.editGroupPost(groupId: java.util.UUID)
+POST    /groups                         controllers.GroupController.addGroup
+GET     /groups/:groupId                controllers.GroupController.editGroup(groupId: java.util.UUID)
+POST    /groups/:groupId                controllers.GroupController.editGroupPost(groupId: java.util.UUID)
 GET     /groups/:groupId/users          controllers.UserController.add(groupId: java.util.UUID)
 POST    /groups/:groupId/users          controllers.UserController.addPost(groupId: java.util.UUID)
 GET     /group/delete/unused/:groupId   controllers.GroupController.deleteUnusedGroupById(groupId: java.util.UUID)
